server/requirements.txt
sentence-transformers==5.1.0
accelerate==1.10.1
websockets==15.0.1
chromadb==1.0.21
python-dotenv==1.1.1
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
transformers==4.46.3
opencv-python-headless==4.12.0.88
Pillow==11.3.0
huggingface_hub==0.34.4
timm==1.0.19
decord==0.6.0
ffmpeg==1.4
imageio==2.37.0
imageio-ffmpeg==0.6.0
ffmpeg-python==0.2.0
librosa==0.11.0
soundfile==0.13.1
sacrebleu==2.5.1
nemo_toolkit[asr,tts] @ git+https://github.com/NVIDIA/NeMo.git
